Move tests of Next pages
