+++ packages/web-client/src/__tests__/pages/_error.spec.tsx
@@ -1,7 +1,7 @@
 import { cleanup } from '@testing-library/react';
 import React from 'react';
 
-import Error from '../_error';
+import Error from '../../pages/_error';
 import { render } from '../../tests/render';
 
 describe('Error', () => {
